A program that launches other programs
No seriously, we needed this. Once you launch this program with admin
credentials, every other program it launches also loads with admin
credentials. My companies policy says that we are not to log in to
company assets with our admin accounts. Also we have to use a 2 facter
authentication in order to get an admin password to copy paste in the
first place. The solution? use Explorer++ (which is buggy as can be) or
a company app called cprocess that only has 6 buttons to use and
requires a text input from an ini file, and I'm not even sure if it
takes in parameters. Cprocess+ (my app) can take in as many programs as
your RAM can handle, and since it just uses strings really, it's a lot.
Quick solution was a winforms app that used the default libraries and
xml serializer to save and load 'profiles' if desired, but I have a WPF
version coming out soon that may just use a binary serializer just to be
even more lightweight.